# apis/openapi.yaml
openapi: 3.1.0
info:
  title: QAForge AI API
  version: 0.1.0
  description: |
    Control Plane API for QAForge AI. Implements the surface defined in
    PRD §13. Phase 0 publishes the contract; endpoints are stubs and will
    fill in across Phase 1 and Phase 2.
  contact:
    name: QAForge Engineering
servers:
  - url: http://localhost:8000
    description: Local
  - url: https://api.staging.qaforge.example.com
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/workspaces:
    get:
      tags: [workspaces]
      summary: List workspaces in the caller's tenant
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: include_archived, in: query, schema: { type: boolean, default: false } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [workspaces]
      summary: Create a workspace
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /api/v1/workspaces/{workspace_id}/environments:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/TenantHeader"
      - $ref: "#/components/parameters/UserHeader"
    get:
      tags: [workspaces]
      summary: List environments configured on a workspace
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvironmentList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/workspaces/{workspace_id}/environments/{name}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/TenantHeader"
      - $ref: "#/components/parameters/UserHeader"
      - { name: name, in: path, required: true, schema: { type: string, pattern: "^[a-zA-Z0-9_-]+$", maxLength: 50 } }
    put:
      tags: [workspaces]
      summary: Create or update an environment by name
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnvironmentUpsert" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Environment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
    get:
      tags: [workspaces]
      summary: Get a workspace environment by name
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Environment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [workspaces]
      summary: Delete a workspace environment
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/model-registry/{entry_id}/scorecard:
    parameters:
      - name: entry_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [model-registry]
      summary: Attach an eval scorecard to a candidate
      responses:
        "200": { description: OK }
        "404": { description: Not found }
  /api/v1/model-registry/{entry_id}/decision:
    parameters:
      - name: entry_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [model-registry]
      summary: Record go/no-go on a candidate
      responses:
        "200": { description: OK }
        "404": { description: Not found }
        "409": { description: Conflict — invalid lifecycle transition }
  /api/v1/model-registry/{entry_id}/deprecate:
    parameters:
      - name: entry_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [model-registry]
      summary: Schedule sunset for a previously-pinned model
      responses:
        "200": { description: OK }
        "404": { description: Not found }
        "409": { description: Conflict — model is not pinned }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    TenantHeader:
      name: X-QAForge-Tenant-Id
      in: header
      required: true
      description: |
        Caller's tenant id. Phase 1 reads this from a header; Phase 3
        (Story 3.1.3) replaces the header source with verified OIDC claims
        without changing the request shape.
      schema: { type: string, format: uuid }
    UserHeader:
      name: X-QAForge-User-Id
      in: header
      required: false
      description: Caller's user id (optional in Phase 1).
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid tenant identity
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Resource conflict (e.g. duplicate name within tenant)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
    ApplicationType:
      type: string
      enum:
        - web_api
        - web_ui
        - web_full_stack
        - mobile_app
        - backend_service
        - library
        - other
    WorkspaceCreate:
      type: object
      required: [name, application_type]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        application_type: { $ref: "#/components/schemas/ApplicationType" }
        repo_url: { type: string, maxLength: 500, nullable: true }
        default_branch: { type: string, default: main, minLength: 1, maxLength: 100 }
        environments:
          type: array
          items: { type: string }
          uniqueItems: true
        description: { type: string, maxLength: 1000, nullable: true }
...
    Workspace:
      type: object
      required:
        - id
        - tenant_id
        - name
        - default_branch
        - application_type
        - environments
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        name: { type: string }
        repo_url: { type: string, nullable: true }
        default_branch: { type: string }
        application_type: { $ref: "#/components/schemas/ApplicationType" }
        environments:
          type: array
          items: { type: string }
        description: { type: string, nullable: true }
        created_by: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        archived_at: { type: string, format: date-time, nullable: true }
    WorkspaceList:
      type: object
      required: [workspaces, limit, offset]
      properties:
        workspaces:
          type: array
          items: { $ref: "#/components/schemas/Workspace" }
        limit: { type: integer }
        offset: { type: integer }
...
    EnvironmentUpsert:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string, pattern: "^[a-zA-Z0-9_-]+$", maxLength: 50 }
        base_url: { type: string, maxLength: 500, nullable: true }
        is_production: { type: boolean, default: false }
        variables:
          type: object
          additionalProperties: true
        description: { type: string, maxLength: 500, nullable: true }
    Environment:
      type: object
      required: [id, workspace_id, name, is_production, variables, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        name: { type: string }
        base_url: { type: string, nullable: true }
        is_production: { type: boolean }
        variables:
          type: object
          additionalProperties: true
        description: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    EnvironmentList:
      type: object
      required: [environments]
      properties:
        environments:
          type: array
          items: { $ref: "#/components/schemas/Environment" }
...